tests fixed

+++ yatube/posts/views.py
@@ -135,6 +135,8 @@ def follow_index(request):
 @login_required
 def profile_follow(request, username):
     author = get_object_or_404(User, username=username)
+    if author == request.user:
+        return redirect(reverse('posts:profile', kwargs={'username': username}))
     if not Follow.objects.filter(user=request.user, author=author).exists():
         Follow.objects.create(user=request.user, author=author)
 
@@ -144,6 +146,8 @@ def profile_follow(request, username):
 @login_required
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
+    if author == request.user:
+        return redirect(reverse('posts:profile', kwargs={'username': username}))
     follow = get_object_or_404(Follow, user=request.user, author=author)
     follow.delete()
